% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{f2022}
\alias{f2022}
\title{Dataset Description}
\format{
A data frame with 483 rows and 18 variables:
\describe{
\item{\code{No}}{Numeric vector representing the observation number.}
\item{\code{Tube_No}}{Numeric vector representing the tube number.}
\item{\code{Species_one}}{Character vector representing the first species.}
\item{\code{Species_two...4}}{Character vector representing the second species.}
\item{\code{Block}}{Character vector representing the block.}
\item{\code{Location}}{Character vector representing the location.}
\item{\code{Treatment}}{Character vector representing the treatment.}
\item{\code{Planting_date}}{Character vector representing the planting date.}
\item{\code{Surv_final}}{Numeric vector representing the final survival status.}
\item{\code{sp1_Abovgd_mass}}{Numeric vector representing the aboveground mass for species one.}
\item{\code{Community_root_mass}}{Numeric vector representing the community root mass.}
\item{\code{sp1_Total_Mass}}{Numeric vector representing the total mass for species one.}
\item{\code{Comm_RMF}}{Numeric vector representing the community root mass fraction.}
\item{\code{No_of_Plants}}{Numeric vector representing the number of plants.}
\item{\code{Control, Intra, Inter}}{Character vector representing the control, intra, or inter status.}
\item{\code{Species_two...16}}{Character vector representing the second species (continued).}
\item{\code{sp2_Abovgd_mass}}{Numeric vector representing the aboveground mass for species two.}
\item{\code{sp2_Total_Mass}}{Numeric vector representing the total mass for species two.}
}
}
\source{
\link{https://sciendo.com/article/10.2478/foecol-2022-0019}
}
\usage{
f2022
}
\description{
This dataset contains information from the native herbaceous competition dataset.
}
\details{
Abstract: Understanding cases in restoration and in agriculture in which species diversity improves productivity and ecosystem functioning is crucial due the need to restore degraded habitat and improve crop productivity for a growing human population. Reaching these diversity benefits is likely influenced by the dynamic of less negative interspecific than intraspecific interactions that promote diversity. But further testing is needed to understand the relationship of intraspecific-relative to interspecific interactions. Here I used seedlings from three native and one introduced species used in restoration in the western United States in pairwise interaction combinations and found that the study species varied in shoot biomass in response to interaction treatments of the control, intraspecific, and interspecific interactions (R2 = 0.7, p < 0.001), and that intraspecific interactions were more negative than interspecific interactions for four of five of the pairings. Overall, as shoot mass size differences increased between interspecific neighbors, interactions became more positive (R2 = 0.6, p < 0.001). These findings point to variability in species responses in whether the focal species compete more intensely with conspecific or heterospecific neighbors and indicates the need for more careful selection of interacting species for meeting both agricultural and restoration goals.
}
\examples{
\dontrun{
  head(f2022)
}
}
\keyword{datasets}
