% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{ffk2024}
\alias{ffk2024}
\title{Dataset Description}
\format{
A data frame with 616 observations and 41 variables:
\describe{
\item{\code{X}}{Integer. Observation number.}
\item{\code{No.}}{Integer. Sample number.}
\item{\code{Block_No}}{Integer. Block number.}
\item{\code{Focal_species}}{Character. Species of interest.}
\item{\code{Species_code}}{Character. Species code.}
\item{\code{Population_code}}{Character. Population code.}
\item{\code{Treatment}}{Character. Type of treatment.}
\item{\code{Comp_trt}}{Character. Competitor treatment.}
\item{\code{Competitor}}{Character. Competitor species.}
\item{\code{Time_in_Exp}}{Integer. Time in the experiment.}
\item{\code{initial_nonfocal_leaf_count}}{Integer. Initial leaf count for non-focal species.}
\item{\code{initial_focal_leaf_count}}{Integer. Initial leaf count for focal species.}
\item{\code{nonfocal_root_length}}{Numeric. Root length for non-focal species.}
\item{\code{nonfocal_shoot_height}}{Numeric. Shoot height for non-focal species.}
\item{\code{nonfocal_root_mass}}{Numeric. Root mass for non-focal species.}
\item{\code{nonfocal_shoot_mass}}{Numeric. Shoot mass for non-focal species.}
\item{\code{focal_root_length}}{Numeric. Root length for focal species.}
\item{\code{focal_shoot_height}}{Numeric. Shoot height for focal species.}
\item{\code{focal_root_mass}}{Numeric. Root mass for focal species.}
\item{\code{focal_shoot_mass}}{Numeric. Shoot mass for focal species.}
\item{\code{Trt}}{Character. Treatment details.}
\item{\code{focal_species}}{Character. Focal species name.}
\item{\code{Type}}{Character. Type of plant.}
\item{\code{nonfocal_leaf_count}}{Integer. Leaf count for non-focal species.}
\item{\code{focal_leaf_count}}{Integer. Leaf count for focal species.}
\item{\code{focal_full_mass}}{Numeric. Full mass for focal species.}
\item{\code{nonfocal_full_mass}}{Numeric. Full mass for non-focal species.}
\item{\code{focal_rmf}}{Numeric. Root mass fraction for focal species.}
\item{\code{nonfocal_rmf}}{Numeric. Root mass fraction for non-focal species.}
\item{\code{rmf_diff}}{Numeric. Difference in root mass fraction.}
\item{\code{shoot_height_diff}}{Numeric. Difference in shoot height.}
\item{\code{root_length_diff}}{Numeric. Difference in root length.}
\item{\code{leaf_count_diff}}{Integer. Difference in leaf count.}
\item{\code{pot_mass}}{Numeric. Pot mass.}
\item{\code{abs_rmf_diff}}{Numeric. Absolute difference in root mass fraction.}
\item{\code{abs_root_length_diff}}{Numeric. Absolute difference in root length.}
\item{\code{abs_shoot_height_diff}}{Numeric. Absolute difference in shoot height.}
\item{\code{final_leaf_diff}}{Integer. Final leaf count difference.}
\item{\code{abs_final_leaf_diff}}{Integer. Absolute final leaf count difference.}
\item{\code{initial_leaf_diff}}{Integer. Initial leaf count difference.}
\item{\code{abs_initial_leaf_diff}}{Integer. Absolute initial leaf count difference.}
}
}
\usage{
ffk2024
}
\description{
This dataset contains information from a plant competition experiment with Heterotheca villosa and Dieteria canescens.
}
\details{
Abstract: Evaluating restoration practices such as admixture provenancing, or the mixing of multiple source populations for greater intraspecific variation in plant restorations, helps to understand the utility of the practice in different scenarios. For admixture provenancing, we do not know whether just any population mixture will deliver positive outcomes in mixed scenarios. To understand how populations and species identity influence plant performance and trait outcomes in mixtures, we assessed trait variation among populations of two forb species native to the western US (Dieteria canescens and Heterotheca villosa) following three interaction treatments (single-population monoculture, two-population mixture, and two-species mixture) with Dieteria canescens as a focal plant in a greenhouse common environment. Plants were harvested after seven weeks of growth and plant height, leaf count, root length, and root mass fraction traits were measured, as well as whole plant biomass. We found that four of six population pairings led to more positive interactions compared to single population monocultures. Neighbor trait differences in shoot height and root length best predicted plant growth following interactions, while increasing neighbor trait differences resulted in positive growth outcomes for all but one trait by treatment interaction. In this single case, plant growth decreased with increasing differences in root mass fraction for the two-species mixture treatment, surprisingly. These findings show nuance in the outcomes of interactions within and between populations and species: they are highly dependent on population identity and traits measured, which are aspects that must be considered when evaluating restoration materials.
}
\examples{
\dontrun{
  head(ffk2024)
}
}
\keyword{datasets}
