% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{fk2020a}
\alias{fk2020a}
\title{Dataset Description}
\format{
A data frame with 520 rows and 8 variables:
\describe{
\item{\code{Block}}{Numeric vector representing the block number.}
\item{\code{TubeNo}}{Numeric vector representing the tube number.}
\item{\code{Accession}}{Character vector representing the accession ID.}
\item{\code{Nutrient}}{Character vector representing the nutrient information.}
\item{\code{ColdStrat}}{Character vector representing the cold stratification date.}
\item{\code{GermDate}}{Character vector representing the germination date.}
\item{\code{ShootMass}}{Numeric vector representing the shoot mass.}
\item{\code{RootMass}}{Character vector representing the root mass.}
}
}
\source{
\link{Provide the source of your dataset, if applicable}
}
\usage{
fk2020a
}
\description{
This dataset contains information from the fk2020a dataset.
}
\examples{
\dontrun{
  head(fk2020a)
}
}
\keyword{datasets}
