% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{fk2020b}
\alias{fk2020b}
\title{Dataset Description}
\format{
A data frame with 180 rows and 15 variables:
\describe{
\item{\code{No.}}{Numeric vector representing the observation number.}
\item{\code{Species}}{Character vector representing the species name.}
\item{\code{Accession}}{Character vector representing the accession ID.}
\item{\code{Source_Accession_Info}}{Character vector representing information about the source accession.}
\item{\code{Collection_type}}{Character vector representing the collection type.}
\item{\code{Block}}{Character vector representing the block information.}
\item{\code{Date_Planted}}{Character vector representing the date planted.}
\item{\code{Date_Harvested}}{Character vector representing the date harvested.}
\item{\code{Age_at_harvest}}{Numeric vector representing the age at harvest.}
\item{\code{Root_length_.cm.}}{Numeric vector representing the root length in centimeters.}
\item{\code{Height_.mm.}}{Numeric vector representing the height in millimeters.}
\item{\code{No_of_leaves}}{Numeric vector representing the number of leaves.}
\item{\code{Root_mass_.mg.}}{Numeric vector representing the root mass in milligrams.}
\item{\code{Shoot_mass_.mg.}}{Numeric vector representing the shoot mass in milligrams.}
\item{\code{No_of_Root_tips}}{Numeric vector representing the number of root tips.}
}
}
\source{
https://onlinelibrary.wiley.com/doi/abs/10.1111/rec.13336
}
\usage{
fk2020b
}
\description{
This dataset contains information from the Pseudoroegneria spicata dataset:
}
\details{
Abstract: Restoration practitioners have many seed material choices when restoring plant communities, and cultivars may be the most affordable and accessible material available for some species. However, the process of plant selection and commercial seed production can limit trait variability critical to survival and persistence in heterogeneous environments. Several studies show the impacts of plant selection and commercial production on trait values and variation in aboveground traits. Yet, researchers rarely assess impacts on root traits in wild-collected material relative to cultivars. This is a critical gap, especially in arid environments where root traits play key roles in plant survival. We grew Pseudoroegneria spicata seedlings from three wild accessions and three cultivars (the cultivars “Whitmar” and “Goldar,” and the “selected germplasm polycross” “P-7”) to compare root and shoot traits and variability. We grew seedlings in a growth chamber for 4 weeks in sand, which had low moisture content and few added nutrients, resulting in stressful conditions. We harvested the plants and compared trait variation and averages between wild-collected versus cultivars for four shoot and four root traits. We found that wild accessions had marginally greater variation in root tip count, root mass fraction, and shoot mass fraction. Importantly, wild-collected plants had 51\% more root tips on average, greater variability in four of seven traits that exhibited differing variability between collection types, and greater survival compared to cultivars. These results indicate differences in traits and variation in wild and cultivated accessions that should be considered when choosing material for restoration use.
}
\examples{
\dontrun{
  head(fk2020b)
}
}
\keyword{datasets}
