% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{fvk2024}
\alias{fvk2024}
\title{Foxx Disrupting Ecology Lab data package
Dataset Description}
\format{
A data frame with 432 rows and 28 variables:
\describe{
\item{\code{...1}}{Numeric vector representing the first column.}
\item{\code{sample}}{Numeric vector representing the sample number.}
\item{\code{site}}{Character vector representing the site name.}
\item{\code{code}}{Character vector representing the code.}
\item{\code{invasive_exposure}}{Character vector representing invasive exposure (yes/no).}
\item{\code{generation}}{Character vector representing the generation.}
\item{\code{maternal_line}}{Character vector representing the maternal line.}
\item{\code{block}}{Numeric vector representing the block.}
\item{\code{shoot_mass}}{Numeric vector representing the shoot mass.}
\item{\code{root_mass}}{Numeric vector representing the root mass.}
\item{\code{height}}{Numeric vector representing the height.}
\item{\code{leaf_count}}{Numeric vector representing the leaf count.}
\item{\code{leaf_width}}{Numeric vector representing the leaf width.}
\item{\code{tillers}}{Numeric vector representing the tillers.}
\item{\code{fibrous_root_count}}{Numeric vector representing the fibrous root count.}
\item{\code{longest_fibrous_root}}{Numeric vector representing the longest fibrous root.}
\item{\code{nonfibrous_root_count}}{Numeric vector representing the nonfibrous root count.}
\item{\code{longest_nonfibrous_root}}{Numeric vector representing the longest nonfibrous root.}
\item{\code{trt}}{Character vector representing the treatment.}
\item{\code{mat_line}}{Character vector representing the maternal line.}
\item{\code{root_length}}{Numeric vector representing the root length.}
\item{\code{total_mass}}{Numeric vector representing the total mass.}
\item{\code{rmf}}{Numeric vector representing the RMF.}
\item{\code{primary_root_count}}{Numeric vector representing the primary root count.}
}
}
\source{
\link{Provide the source of your dataset, if applicable}
}
\usage{
fvk2024
}
\description{
This dataset contains information from the Sporobolus airoides project:
}
\details{
Abstract: Exposure to competition from invasive plants over multiple generations, or multigenerational invasive exposure can drive rapid native plant trait change. While invasive plant effects can be concentrated belowground, few studies consider native plant root trait responses to multigenerational invasive exposure.  So here we quantified root and shoot trait responses of the native grass Sporobolus airoides in response to multigenerational invasive exposure. S. airoides was sourced from invader-experienced subpopulations that co-occurred in the field for 58 years on average with invasive Russian knapweed (Rhaponticum repens) and from nearby invader-naïve plants that did not co-occur with the invasive. Maternal plants of these subpopulations were collected in, and transplanted to a greenhouse in Wyoming, USA. The first generation of seeds from these transplants were collected and grown for another round of seed collection (generation two). We grew invader-experienced and invader-naïve seedlings from both generations in a growth chamber in Illinois, USA. We found the largest shifts in root traits with invasive exposure status (e.g., for root mass fraction (standardized mean difference (SMD) = 0.5), primary root count (SMD) = 0.52), root length (SMD) = 0.44)); invader-experienced plants had significantly larger root traits in generation two, which were associated with greater survival, and thus may be adaptive.  Since invader-experienced accessions have been evaluated as valuable restoration material in invaded sites, these findings highlight the need for continued root research in multigenerational invasive plant interactions and the need to evaluate more than one generation removed from a biotic influence.
}
\examples{
\dontrun{
  head(fvk2024)
}
}
\keyword{datasets}
