% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{fw2021}
\alias{fw2021}
\title{Dataset Description}
\format{
A data frame with 78 observations and 12 variables:
\describe{
\item{No.}{Numeric. Observation number.}
\item{Tube}{Numeric. Tube number.}
\item{Species}{Character. Species of the plant.}
\item{Accession}{Character. Accession information.}
\item{Soil.type}{Character. Type of soil.}
\item{Tray}{Character. Tray identifier.}
\item{Date.planted}{Character. Date when the plant was planted.}
\item{leaf.lenght.7d}{Numeric. Length of the leaf after 7 days (in millimeters).}
\item{LR}{Integer. Presence of lateral roots (0 = absent, 1 = present).}
\item{MR}{Integer. Presence of main roots (0 = absent, 1 = present).}
\item{Root.tips}{Integer. Number of root tips observed.}
\item{Notes}{Character. Additional notes or comments.}
}
}
\source{
\link{https://sciendo.com/article/10.2478/foecol-2021-0020}
}
\usage{
fw2021
}
\description{
This dataset contains information from the Bromus tectorum study.
}
\details{
Abstract: The relationships between shoot and root traits can inform plant selection for restoration, forestry, and agriculture and help to identify relationships that inform plant productivity and enhance their performance. But the strength of coordination between above- and belowground morphological and physiological traits varies due to differences in edaphic properties and population variation. More assessments are needed to determine what conditions influence these relationships. So, we tested whether plant population and soil texture affect the relationship between shoot and root traits which have important ecological ramifications for competition and resource capture: shoot height and root tip production. We grew seedlings of two populations of Bromus tectorum due to is fast growing nature in a growth chamber in loam soil, sand, and clay. We found variation in height by plant population and the substrate used (R2 = 0.44, p < 0.0001), and variation in root tip production by the substrate used (R2 = 0.33, p < 0.0001). Importantly, we found that relationships between shoot height and root tip production varied by soil texture and population (R2 = 0.54, p < 0.0001), and growth in sand produced the strongest relationship and was the most water deficient substrate (R2 = 0.32). This shows that screening populations under several environments influences appropriate plant selection.
}
\examples{
\dontrun{
  head(fw2021)
}
}
\keyword{datasets}
