% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{f2022}
\alias{f2022}
\title{Dataset Description}
\format{
A data frame with 483 rows and 18 variables:
\describe{
\item{\code{No}}{Numeric vector representing the observation number.}
\item{\code{Tube_No}}{Numeric vector representing the tube number.}
\item{\code{Species_one}}{Character vector representing the first species.}
\item{\code{Species_two...4}}{Character vector representing the second species.}
\item{\code{Block}}{Character vector representing the block.}
\item{\code{Location}}{Character vector representing the location.}
\item{\code{Treatment}}{Character vector representing the treatment.}
\item{\code{Planting_date}}{Character vector representing the planting date.}
\item{\code{Surv_final}}{Numeric vector representing the final survival status.}
\item{\code{sp1_Abovgd_mass}}{Numeric vector representing the aboveground mass for species one.}
\item{\code{Community_root_mass}}{Numeric vector representing the community root mass.}
\item{\code{sp1_Total_Mass}}{Numeric vector representing the total mass for species one.}
\item{\code{Comm_RMF}}{Numeric vector representing the community root mass fraction.}
\item{\code{No_of_Plants}}{Numeric vector representing the number of plants.}
\item{\code{Control, Intra, Inter}}{Character vector representing the control, intra, or inter status.}
\item{\code{Species_two...16}}{Character vector representing the second species (continued).}
\item{\code{sp2_Abovgd_mass}}{Numeric vector representing the aboveground mass for species two.}
\item{\code{sp2_Total_Mass}}{Numeric vector representing the total mass for species two.}
}
}
\source{
\link{https://sciendo.com/article/10.2478/foecol-2022-0019}
}
\usage{
f2022
}
\description{
This dataset contains information from the f2022 dataset.
}
\examples{
\dontrun{
  head(f2022)
}
}
\keyword{datasets}
