% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{fk2020b}
\alias{fk2020b}
\title{Dataset Description}
\format{
A data frame with 180 rows and 15 variables:
\describe{
\item{\code{No.}}{Numeric vector representing the observation number.}
\item{\code{Species}}{Character vector representing the species name.}
\item{\code{Accession}}{Character vector representing the accession ID.}
\item{\code{Source_Accession_Info}}{Character vector representing information about the source accession.}
\item{\code{Collection_type}}{Character vector representing the collection type.}
\item{\code{Block}}{Character vector representing the block information.}
\item{\code{Date_Planted}}{Character vector representing the date planted.}
\item{\code{Date_Harvested}}{Character vector representing the date harvested.}
\item{\code{Age_at_harvest}}{Numeric vector representing the age at harvest.}
\item{\code{Root_length_.cm.}}{Numeric vector representing the root length in centimeters.}
\item{\code{Height_.mm.}}{Numeric vector representing the height in millimeters.}
\item{\code{No_of_leaves}}{Numeric vector representing the number of leaves.}
\item{\code{Root_mass_.mg.}}{Numeric vector representing the root mass in milligrams.}
\item{\code{Shoot_mass_.mg.}}{Numeric vector representing the shoot mass in milligrams.}
\item{\code{No_of_Root_tips}}{Numeric vector representing the number of root tips.}
}
}
\source{
https://onlinelibrary.wiley.com/doi/abs/10.1111/rec.13336
}
\usage{
fk2020b
}
\description{
This dataset contains information about from the fk2020b publication.
}
\examples{
\dontrun{
  head(fk2020b)
}
}
\keyword{datasets}
