% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foxxlab-package.R
\docType{data}
\name{fvk2024}
\alias{fvk2024}
\title{Foxx Disrupting Ecology Lab data package
Dataset Description}
\format{
A data frame with 432 rows and 28 variables:
\describe{
\item{\code{...1}}{Numeric vector representing the first column.}
\item{\code{sample}}{Numeric vector representing the sample number.}
\item{\code{site}}{Character vector representing the site name.}
\item{\code{code}}{Character vector representing the code.}
\item{\code{invasive_exposure}}{Character vector representing invasive exposure (yes/no).}
\item{\code{generation}}{Character vector representing the generation.}
\item{\code{maternal_line}}{Character vector representing the maternal line.}
\item{\code{block}}{Numeric vector representing the block.}
\item{\code{shoot_mass}}{Numeric vector representing the shoot mass.}
\item{\code{root_mass}}{Numeric vector representing the root mass.}
\item{\code{height}}{Numeric vector representing the height.}
\item{\code{leaf_count}}{Numeric vector representing the leaf count.}
\item{\code{leaf_width}}{Numeric vector representing the leaf width.}
\item{\code{tillers}}{Numeric vector representing the tillers.}
\item{\code{fibrous_root_count}}{Numeric vector representing the fibrous root count.}
\item{\code{longest_fibrous_root}}{Numeric vector representing the longest fibrous root.}
\item{\code{nonfibrous_root_count}}{Numeric vector representing the nonfibrous root count.}
\item{\code{longest_nonfibrous_root}}{Numeric vector representing the longest nonfibrous root.}
\item{\code{trt}}{Character vector representing the treatment.}
\item{\code{mat_line}}{Character vector representing the maternal line.}
\item{\code{root_length}}{Numeric vector representing the root length.}
\item{\code{total_mass}}{Numeric vector representing the total mass.}
\item{\code{rmf}}{Numeric vector representing the RMF.}
\item{\code{primary_root_count}}{Numeric vector representing the primary root count.}
\item{\code{sqrt_root_mass}}{Numeric vector representing the square root of root mass.}
\item{\code{sqrt_shoot_mass}}{Numeric vector representing the square root of shoot mass.}
\item{\code{asin_leaf_width}}{Numeric vector representing the arcsine of leaf width.}
\item{\code{sqrt_total_mass}}{Numeric vector representing the square root of total mass.}
}
}
\source{
\link{Provide the source of your dataset, if applicable}
}
\usage{
fvk2024
}
\description{
This dataset contains information from the fvk2024 manuscript.
}
\examples{
\dontrun{
  head(fvk2024)
}
}
\keyword{datasets}
